Transformation of structure: Login page with authentication, new homepage to replace old exact '/' path, following new Balsamiq wireframe design
